microservices/processor/requirements.txt
google-cloud-pubsub==2.18.1
google-auth==2.27.0

psycopg2-binary==2.9.9

python-dotenv==1.0.1
